fix [#117] 애플게임센터 로그아웃됐을 때, 앱 접속시 무한로딩 버그 fix
