+++ FullStackCodingBot/FullStackCodingBot/View/Controller/ViewController/Main/MainViewController.swift
@@ -31,21 +31,13 @@ final class MainViewController: UIViewController, ViewModelBindableType {
         buttonController.bind { [unowned self] viewController in
             self.viewModel.makeMoveAction(to: viewController)
         }
-        
-        viewModel.firebaseDidLoad
-            .subscribe(on: MainScheduler.asyncInstance)
-            .subscribe(onNext: { [unowned self] isLoaded in
-                if !isLoaded {
-                    self.viewModel.startLoading()
-                }
-            }).disposed(by: rx.disposeBag)
     }
 }
 
+// MARK: Setup
 private extension MainViewController {
     
     private func setup() {
-        setupAppleGameCenterLogin()
         setupTitleLabel()
         titleLabel.startTypewritingAnimation()
     }
@@ -55,34 +47,5 @@ private extension MainViewController {
         let attributedString = NSMutableAttributedString(string: Text.title)
         attributedString.addAttribute(.font, value: font, range: .init(location: 0, length: Text.title.count))
         titleLabel.attributedText = attributedString
-    }
-    
-    private func setupAppleGameCenterLogin() {
-        GKLocalPlayer.local.authenticateHandler = { gcViewController, error in
-            guard error == nil else { return }
-            
-            if GKLocalPlayer.local.isAuthenticated {
-                GameCenterAuthProvider.getCredential { credential, error in
-                    guard error == nil else { return }
-                    
-                    Auth.auth().signIn(with: credential!) { [unowned self] user, error in
-                        guard error == nil else { return }
-                        
-                        if user != nil {
-                            self.viewModel.getUserInformation()
-                        }
-                    }
-                }
-            } else if let gcViewController = gcViewController {
-                print(gcViewController)
-            }
-        }
-    }
-}
-
-extension MainViewController: GKGameCenterControllerDelegate {
-    
-    func gameCenterViewControllerDidFinish(_ gameCenterViewController: GKGameCenterViewController) {
-        print("GCVC DID FINISHED")
     }
 }
